Add Mininal Versions of bCmd
Minimal is bCmd without settings options and such, still fully functioning as a cmd replacement.
Minimal_Smaller is Minimal but without the Microsoft watermark, fully functions.
Minimal_Smallest is completely butchered. Environmental variables do not work, and the console is spammed as @echo off is never ran. Loads of issues.